requirements.txt
# Core dependencies
pandas>=1.3.0
numpy>=1.20.0
python-dotenv>=0.19.0


# Web Framework (if needed)
flask>=2.0.0

# Data Visualization (if needed)
matplotlib>=3.4.0
seaborn>=0.11.0

# Development Tools
pytest>=6.2.0
black>=21.5b0
flake8>=3.9.0
streamlit>=0.80.0
groq>=0.7.0
fpdf
praw

